Added NeighborCache class
I didn't like the hardcoded templates for nearest neighbor that I have been using, so through an extended discussion with ChatGPT, we developed this tool to generate templates for a variety of shapes that can be used for searching neighbors, among other uses.  Not currently using it in main code, but it's here if I want it or for future reference.